requirements.txt
pandas>=1.1.4
numpy>=1.18.5
scikit-learn
pymongo==4.6.1
fastapi==0.110.2
uvicorn==0.20.0
python-multipart
dnspython==2.2.1
certifi==2019.9.11
gunicorn
langchain==0.3.0
langchain-mongodb==0.2.0
langchain-core 
langchain-openai 
langchain-community
openai
python-dotenv==1.0.0
pinecone-client==2.2.4
chromadb
faiss-cpu==1.7.4
pycryptodome==3.15.0
openapi-schema-pydantic==1.2.4
tiktoken
python-multipart
yagmail[all]
azure-storage-blob
uuid
streamlit_chat
bcrypt